tests for summary_stats_display
