+++ tests/pluggable_analysis_framework_test.py
@@ -11,6 +11,8 @@ from buckaroo.analysis_management import (
 
 from buckaroo.analysis import (TypingStats, DefaultSummaryStats)
 
+from .fixtures import (test_df, df, DistinctCount, Len, DistinctPer, DCLen, DependsNoProvides)
+
 class NoRoute(ColAnalysis):    
     provided_summary = ['not_used']
     requires_summary = ["does_not_exist"]
@@ -22,7 +24,6 @@ class CycleA(ColAnalysis):
 class CycleB(ColAnalysis):
     provided_summary = ['cycle_b']
     requires_summary = ["cycle_a"]
-
 
 class CA_AB(ColAnalysis):
     provided_summary = ["a", "b"]
@@ -37,48 +38,6 @@ class CA_EF(ColAnalysis):
 class CA_G(ColAnalysis):
     provided_summary = ["g"]
     requires_summary = ["e"]
-
-test_df = pd.DataFrame({
-        'normal_int_series' : pd.Series([1,2,3,4]),
-        'empty_na_ser' : pd.Series([], dtype="Int64"),
-        'float_nan_ser' : pd.Series([3.5, np.nan, 4.8])
-    })
-
-class DistinctCount(ColAnalysis):
-    requires_raw = True
-    provided_summary = ["distinct_count"]
-    @staticmethod
-    def summary(sampled_ser, summary_ser, raw_ser):
-        val_counts = raw_ser.value_counts()
-        distinct_count= len(val_counts)
-        return {'distinct_count': distinct_count}
-
-class Len(ColAnalysis):
-    provided_summary = ["len"]
-    requires_raw = True
-    @staticmethod
-    def summary(sampled_ser, summary_ser, raw_ser):
-        return {'len': len(raw_ser)}
-
-class DCLen(ColAnalysis):
-    provided_summary = ["len", "distinct_count"]
-    requires_raw = True
-    @staticmethod
-    def summary(sampled_ser, summary_ser, raw_ser):
-        val_counts = raw_ser.value_counts()
-        distinct_count= len(val_counts)
-        return {'len':len(raw_ser), 'distinct_count':distinct_count}
-
-class DistinctPer(ColAnalysis):
-    provided_summary = ["distinct_per"]
-    requires_summary = ["len", "distinct_count"]
-    
-    @staticmethod
-    def summary(sampled_ser, summary_ser, raw_ser):
-        return {'distinct_per': summary_ser.loc['distinct_count'] / summary_ser.loc['len']}
-
-class DependsNoProvides(ColAnalysis):
-    requires_summary = ["len"]
 
 
 class TestOrderAnalysis(unittest.TestCase):
@@ -119,63 +78,6 @@ class TestOrderAnalysis(unittest.TestCase):
         with self.assertRaises(NotProvidedException):
             check_solvable([NoRoute])
 
-df = pd.read_csv('./examples/data/2014-01-citibike-tripdata.csv')
-
-class TestAnalysisPipeline(unittest.TestCase):
-    def test_produce_summary_df(self):
-        produce_summary_df(test_df, [DistinctCount, Len, DistinctPer], 'test_df')
-
-
-    def test_pipeline_base(self):
-        ap = AnalsysisPipeline([TypingStats, DefaultSummaryStats])
-        #just verify that there are no errors
-        ap.process_df(df)
-
-    def test_add_aobj(self):
-        ap = AnalsysisPipeline([TypingStats, DefaultSummaryStats])
-        class Foo(ColAnalysis):
-            provided_summary = [
-                'foo',]
-            requires_summary = ['length']
-
-            @staticmethod
-            def summary(sampled_ser, summary_ser, ser):
-                return dict(foo=8)
-        ap.add_analysis(Foo)
-        sdf, _unused = ap.process_df(df)
-        self.assertEqual(sdf.loc['foo']['tripduration'], 8)
-
-    def test_replace_aobj(self):
-        ap = AnalsysisPipeline([TypingStats, DefaultSummaryStats])
-        class Foo(ColAnalysis):
-            provided_summary = [
-                'foo',]
-            requires_summary = ['length']
-
-            @staticmethod
-            def summary(sampled_ser, summary_ser, ser):
-                return dict(foo=8)
-        ap.add_analysis(Foo)
-        sdf, _unused = ap.process_df(df)
-        self.assertEqual(sdf.loc['foo']['tripduration'], 8)
-        #18 facts returned about tripduration
-        self.assertEqual(len(sdf['tripduration']), 18)
-        #Create an updated Foo that returns 9
-        class Foo(ColAnalysis):
-            provided_summary = [
-                'foo',]
-            requires_summary = ['length']
-
-            @staticmethod
-            def summary(sampled_ser, summary_ser, ser):
-                return dict(foo=9)
-        ap.add_analysis(Foo)
-        sdf2, _unused = ap.process_df(df)
-        self.assertEqual(sdf2.loc['foo']['tripduration'], 9)
-        #still 18 facts returned about tripduration
-        self.assertEqual(len(sdf2['tripduration']), 18)
-        #Create an updated Foo that returns 9
-        
 
 
 """
